chore: example for simple scene about two balls' transformation (#4)
* chore(example): working in rpc transport.

* chore(bundle): bring up new build.

* chore(bundle): test

* chore: test

* chore(husky): init

* chore(husky): commit-msg valid tool

* chore(example): complete mock-net class

* chore(example): ball move by self

* chore: delete index.html

* docs: add explanation about example

+++ src/data/string-databuffer.ts
@@ -1,10 +1,10 @@
 import { IDataBuffer, IDatabufferWriter } from "./serializable";
 
 const tempTypedBuffer = {
-    int: new Int16Array(1),
-    uint: new Uint16Array(1),
-    short: new Int8Array(1),
-    ushort: new Uint8Array(1),
+    int: new Int32Array(1),
+    uint: new Uint32Array(1),
+    short: new Int16Array(1),
+    ushort: new Uint16Array(1),
     long: new Int32Array(1),
     ulong: new Uint32Array(1),
     float: new Float32Array(1),
